# Conflicts:
#	pom.xml
#	src/main/java/ru/home_work/t1_school/aspect/LogAspect.java
#	src/main/java/ru/home_work/t1_school/controller/TaskControllerImpl.java
#	src/main/java/ru/home_work/t1_school/service/TaskServiceImpl.java

+++ src/main/java/ru/home_work/t1_school/service/TaskServiceImpl.java
@@ -5,6 +5,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import ru.home_work.t1_school.aspect.annotations.LogException;
 import ru.home_work.t1_school.aspect.annotations.LogExecutionTime;
+import ru.home_work.t1_school.aspect.annotations.LogReturning;
 import ru.home_work.t1_school.exception.TaskNotFoundException;
 import ru.home_work.t1_school.model.Task;
 import ru.home_work.t1_school.repository.TaskRepository;
@@ -18,9 +19,10 @@ import java.util.Optional;
 public class TaskServiceImpl implements TaskService {
 
     private final TaskRepository repository;
-
+    private final String TASK_NOT_FOUND_MESSAGE = "Задание с идентификатором %s не найдено";
     @LogExecutionTime
     @LogException
+    @LogReturning
     @Override
     public Task create(Task task) {
         return repository.save(task);
@@ -28,13 +30,14 @@ public class TaskServiceImpl implements TaskService {
 
     @LogExecutionTime
     @LogException
+    @LogReturning
     @Override
     public Task getTask(Long id) {
         Optional<Task> byId = repository.findById(id);
         if (byId.isPresent()) {
             return byId.get();
         } else {
-            String message = String.format("Задание с идентификатором %s не найдено", id);
+            String message = String.format(TASK_NOT_FOUND_MESSAGE, id);
             throw new TaskNotFoundException(message);
         }
     }
@@ -48,7 +51,7 @@ public class TaskServiceImpl implements TaskService {
             task.setId(id);
             repository.save(task);
         } else {
-            String message = String.format("Задание с идентификатором %s не найдено", id);
+            String message = String.format(TASK_NOT_FOUND_MESSAGE, id);
             throw new TaskNotFoundException(message);
         }
 
@@ -58,11 +61,18 @@ public class TaskServiceImpl implements TaskService {
     @LogException
     @Override
     public void delete(Long id) {
-        repository.deleteById(id);
+        Optional<Task> byId = repository.findById(id);
+        if (byId.isPresent()) {
+            repository.deleteById(id);
+        } else {
+            String message = String.format(TASK_NOT_FOUND_MESSAGE, id);
+            throw new TaskNotFoundException(message);
+        }
     }
 
     @LogExecutionTime
     @LogException
+    @LogReturning
     @Override
     public Collection<Task> list() {
         return repository.findAll();
